Better scientist and camera, and Frog player controller
Scientists have paths they follow and will run if they see the frog. Frog is animated and movable via a custom player controller script. Camera collides with walls but is a bit wonky when controlling because of that. Currently the joystick and jump buttons won't work because of the player controller script but can be reenabled by commenting out the #if UNITY_EDITOR_WIN part of the script if you guys want to test the controls. The controls are currently WASD and space but arrows keys work too.
One issue is the vertical doors still block scientists but that's an easy fix.
Another issue is the ceiling may be problematic since the camera collides with things. Scientists can also now catch the frog.