Frog Fix, Camera prefab, and Scientist logic
Frog was missing Animator and Frog script components but is now fixed. Frog was also set to kinematic but is switched off and only using gravity. Player controller script modified to only search for main camera instead of a specific virtual camera. An isometric virtual camera prefab is made. It just needs to have the frog assigned to look at and follow variables. Scientist script modified to allow frog to escape while pick up animation is playing instead of automatically capturing frog when colliding. This means scientist animator is modified to include a trigger parameter and transition from pick up to movement.